Add reducers to remove books from the list by clicking remove button
